backend/requirements.txt

uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
fastapi==0.110.3
starlette==0.37.2
pydantic==2.7.4

